+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/entity/AOPConsumptionNorm.java
@@ -36,40 +36,40 @@ public class AOPConsumptionNorm {
     private UUID materialFkId;
 
     @Column(name = "Jan")
-    private Float jan;
+    private Double jan;
 
     @Column(name = "Feb")
-    private Float feb;
+    private Double feb;
 
     @Column(name = "March")
-    private Float march;
+    private Double march;
 
     @Column(name = "April")
-    private Float april;
+    private Double april;
 
     @Column(name = "May")
-    private Float may;
+    private Double may;
 
     @Column(name = "June")
-    private Float june;
+    private Double june;
 
     @Column(name = "July")
-    private Float july;
+    private Double july;
 
     @Column(name = "Aug")
-    private Float aug;
+    private Double aug;
 
     @Column(name = "Sep")
-    private Float sep;
+    private Double sep;
 
     @Column(name = "Oct")
-    private Float oct;
+    private Double oct;
 
     @Column(name = "Nov")
-    private Float nov;
+    private Double nov;
 
     @Column(name = "Dec")
-    private Float dec;
+    private Double dec;
 
     @Column(name = "AOPYear", length = 100, nullable = false)
     private String aopYear;
