+++ apps/java/libraries/case-engine/src/main/java/com/wks/caseengine/entity/BusinessDemand.java
@@ -28,40 +28,40 @@ public class BusinessDemand {
     private UUID normParameterId;
     
     @Column(name = "Jan")
-    private Float jan;
+    private Double jan;
     
     @Column(name = "Feb")
-    private Float feb;
+    private Double feb;
     
     @Column(name = "March")
-    private Float march;
+    private Double march;
     
     @Column(name = "April")
-    private Float april;
+    private Double april;
     
     @Column(name = "May")
-    private Float may;
+    private Double may;
     
     @Column(name = "June")
-    private Float june;
+    private Double june;
     
     @Column(name = "July")
-    private Float july;
+    private Double july;
     
     @Column(name = "Aug")
-    private Float aug;
+    private Double aug;
     
     @Column(name = "Sep")
-    private Float sep;
+    private Double sep;
     
     @Column(name = "Oct")
-    private Float oct;
+    private Double oct;
     
     @Column(name = "Nov")
-    private Float nov;
+    private Double nov;
     
     @Column(name = "Dec")
-    private Float dec;
+    private Double dec;
     
     @Column(name = "Year", length = 9, nullable = false)
     private String year;
@@ -70,7 +70,7 @@ public class BusinessDemand {
     private UUID plantId;
     
     @Column(name = "AvgTPH")
-    private Float avgTph;
+    private Double avgTph;
     
     @Column(name="Site_FK_Id")
     private UUID siteFKId;
